Improve non deterministic exploration
The exploration tree contains now hard configurations on top and soft
configurations elsewhere. A soft configuration contains only a pointer to
its parent configuraiton in the tree and the associated transition that
should be fired. A hard configuration contains a copy, a snapshot of the
machine. That way the tapes are not copied many times in the nodes of the tree.